feat #1: support time values in graph component

+++ app/src/main/java/org/htwk/pacing/math/Math.kt
@@ -2,14 +2,14 @@ package org.htwk.pacing.math
 
 import androidx.annotation.FloatRange
 
-fun interpolate(a: Float, b: Float, @FloatRange(from = 0.0, to = 1.0)t: Float): Float {
+fun interpolate(a: Double, b: Double, @FloatRange(from = 0.0, to = 1.0) t: Float): Double {
     return a + t * (b - a)
 }
 
-fun remap(value: Float, lowIn: Float, highIn: Float, lowOut: Float, highOut: Float): Float {
+fun remap(value: Double, lowIn: Double, highIn: Double, lowOut: Double, highOut: Double): Double {
     if (lowIn == highIn) {
         println("Error: Invalid input range for remap: lowIn == highIn")
-        return 0f
+        return 0.0
     }
     return lowOut + (value - lowIn) * (highOut - lowOut) / (highIn - lowIn)
 }
